added user logout, login, register logs
